Move more code out of mathics.builtin.__init__
Moved to mathics.eval.builtin. Possibly not the best place, but it is
better and is a start.
