+++ mathics/builtin/box/layout.py
@@ -7,8 +7,6 @@ The routines here assist in boxing at the bottom of the hierarchy. \
 At the other end, the top level, we have a Notebook which is just a \
 collection of Expressions usually contained in boxes.
 """
-# Docs are not yet ready for prime time. Maybe after release 6.0.0.
-no_doc = True
 
 from mathics.builtin.base import Builtin
 from mathics.builtin.box.expression import BoxExpression
@@ -31,6 +29,9 @@ from mathics.core.systemsymbols import (
     SymbolSuperscriptBox,
 )
 from mathics.eval.makeboxes import eval_makeboxes
+
+# Docs are not yet ready for prime time. Maybe after release 6.0.0.
+no_doc = True
 
 
 def to_boxes(x, evaluation: Evaluation, options={}) -> BoxElementMixin:
