+++ mathics/eval/makeboxes.py
@@ -11,7 +11,6 @@ from typing import Any, Dict, Type
 
 from mathics.core.atoms import Complex, Integer, Rational, Real, String, SymbolI
 from mathics.core.convert.expression import to_expression_with_specialization
-from mathics.core.definitions import OutputForms
 from mathics.core.element import BaseElement, BoxElementMixin, EvalMixin
 from mathics.core.evaluation import Evaluation
 from mathics.core.expression import Expression
@@ -36,7 +35,6 @@ from mathics.core.symbols import (
 from mathics.core.systemsymbols import (
     SymbolComplex,
     SymbolMinus,
-    SymbolOutputForm,
     SymbolRational,
     SymbolRowBox,
     SymbolStandardForm,
@@ -130,6 +128,8 @@ def do_format_element(
     Applies formats associated to the expression and removes
     superfluous enclosing formats.
     """
+
+    from mathics.core.definitions import OutputForms
 
     evaluation.inc_recursion_depth()
     try:
